requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.7
django-cors-headers==3.8.0
django-multitenant==2.3.2
django-phonenumber-field==6.0.0
django-rest-swagger==2.2.0
django-tenant-schemas==1.10.0
django-tenants==3.3.2
djangorestframework==3.12.4
idna==3.2
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
mysqlclient==2.1.0
openapi-codec==1.3.2
ordered-set==4.0.2
phonenumbers==8.12.43
psycopg2-binary==2.9.3
PyMySQL==1.0.2
pytz==2021.1
requests==2.26.0
simplejson==3.17.5
six==1.16.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.6
xlrd==1.2.0
